update to 25.1

+++ test/gridengine/test.sh
@@ -10,17 +10,19 @@ docker build --target sge_bioblend_test --tag sge_bioblend_test .
 # galaxy and gridengine:
 EXPORT=`mktemp --directory`
 chmod 777 ${EXPORT}
-docker run --hostname sgemaster --name sgemaster -d -v ${EXPORT}:/export -v $PWD/master_script.sh:/usr/local/bin/master_script.sh sge_master /usr/local/bin/master_script.sh
+docker run -d --rm --hostname sgemaster --name sgemaster -v ${EXPORT}:/export -v $PWD/master_script.sh:/usr/local/bin/master_script.sh sge_master /usr/local/bin/master_script.sh
 # wait for sge master
 sleep 10
 
 # start galaxy
-GALAXY_CONTAINER=quay.io/bgruening/galaxy
+GALAXY_CONTAINER=${GALAXY_CONTAINER:-quay.io/bgruening/galaxy}
+EPHEMERIS_IMAGE=${EPHEMERIS_IMAGE:-quay.io/biocontainers/ephemeris:0.10.11--pyhdfd78af_0}
+GALAXY_WAIT_TIMEOUT=${GALAXY_WAIT_TIMEOUT:-600}
 GALAXY_CONTAINER_NAME=galaxytest
 GALAXY_CONTAINER_HOSTNAME=galaxytest
 GALAXY_ROOT_DIR=/galaxy
 
-docker run -d \
+docker run -d --rm \
            -e SGE_ROOT=/var/lib/gridengine \
            --link sgemaster:sgemaster \
            --name ${GALAXY_CONTAINER_NAME} \
@@ -52,8 +54,13 @@ docker exec sgemaster cat /etc/hosts
 # Add gridengine client host
 echo "Add submit host ${GALAXY_CONTAINER_HOSTNAME}"
 docker exec sgemaster bash -c "qconf -as ${GALAXY_CONTAINER_HOSTNAME}"
-echo "Wait 180sec"
-sleep 180
+echo "Waiting for Galaxy to become ready"
+if ! docker run --rm --link ${GALAXY_CONTAINER_NAME}:galaxytest \
+    ${EPHEMERIS_IMAGE} galaxy-wait -g http://galaxytest --timeout ${GALAXY_WAIT_TIMEOUT}; then
+    echo "Galaxy did not become ready within ${GALAXY_WAIT_TIMEOUT}s."
+    docker logs ${GALAXY_CONTAINER_NAME} || true
+    exit 1
+fi
 
 echo "Exec test"
 docker run --rm --link galaxytest:galaxytest -v $PWD/test_outputhostname.py:/work/test_outputhostname.py sge_bioblend_test python /work/test_outputhostname.py > out
@@ -61,10 +68,8 @@ grep sgemaster out
 RET=$?
 
 # remove container
-docker stop sgemaster
-docker rm sgemaster
-docker stop galaxytest
-docker rm galaxytest
+docker stop sgemaster || true
+docker stop galaxytest || true
 
 # Remove images 
 docker rmi sge_master
